pyproject.toml
[tool.poetry]
name = "watchmate"
version = "0.1.0"
description = ""
authors = ["Behrooz Payoon <behrooz.payoon@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "4.2"
djangorestframework = "^3.15.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
